prototipHistory/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.7.22
Django==4.2.5
django-elasticsearch-dsl==7.2.0
django-elasticsearch-dsl-drf==0.22.1
django-nine==0.2.7
djangorestframework==3.14.0
elastic-transport==8.4.1
elasticsearch==7.14.0
elasticsearch-dsl==7.4.0
Faker==19.6.1
packaging==23.1
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.8.0
urllib3==1.26.16